scripts/requirements.txt
requests>=2.28.0
requests-oauthlib>=1.3.1
python-dotenv>=0.20.0
markdownify>=0.11.0
beautifulsoup4>=4.11.0
